Move public/private key functionality to a new file
